Add a script to build and create a release of Hikari Run ```python python distribute.py ```
to create a distributable build of the engine, including the editor,
cli and necessary files.

Generated artifacts are stored in a folder called `dist`.
If you want a zip file with the contents of this folder to be
generated automatically you can do something like this:
```python
python distribute.py --package=yourfilepath.zip
```