Move fence resetting code to `submit(...)`
The render fence is reset inside the `finish_internal(...)`
method. This causes problems when the Graph is executed in sync mode.
When `render_sync(...)` is called the render fence is reset and is prepared for the next frame.
If `finish_internal(...)` is called right afterwards, we try to
wait on this fence that will never be signaled resulting in an timeout.

To fix this the fence is now reset right before submitting the command buffer.