Fix: VkDevice lost when rendering and presenting at the same time Turns out the Graphics Queue and Present Queue can internally point to the same queue.
This causes crashes when we submit concurrently to both these queues
as VkQueueSubmit must be be externally synchronized per queue.

We fix this by checking if the Graphics Queue and the Present Queue are
indeed the same queue. If they are, during presentation, we submit to
the Graphics Queue which is guarded by a mutex. Otherwise we submit to
the dedicated Present Queue. This doesn't need synchronization as only
the Swapchain can submit to the present queue, and there can be only
one Swapchain per application.